Add guidance for generating an elm client
I saw one of the things on the roadmap was an implementation in Elm. I'm starting to play around with this api for an Elm project and wanted to add the generation command as well as where to find more information for how to create Elm client code from https://api.marvelql.com/. Not sure if this is exactly what you had in mind but figured I'd check in!